+++ CHANGELOG.md
@@ -31,7 +31,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - **CUMULUS-2794**
     - Remove API Collections model and all related test code
     - Remove lambdas/postgres-migration-count-tool, api/endpoints/migrationCounts and api-client/migrationCounts
-    - Remove lambdas/data-migration* tool
+    - Remove lambdas/data-migration1 tool
+    - Remove lambdas/data-migration2 and lambdas/postgres-migration-async-operation
 
 ### Added